speech-api-pipeline/local-pc-resources/requirements.txt
##### Requireements for local pc of speech-api-pipeline #####
boto3
google-cloud
pydub
watchdog
